Add filter reducer
